StreamSPI: do not handle SPI collision
Handle the SPI collision correctly require too much effort. On MUC side we
should store all the incoming frame in a different buffer, and send a special
frame to the main CPU to ask for the missing byte. Recieve the missing byte and
insert in the pending frame, then store the whole frame. But this does not
resolve the problem because a collision can occur while retrieving the missing
byte.

If we do not handle collision AVR recieves a frame with wrong bytes, but the
communication is simple and the CPU can retry with a new frame if it can
recognize the problem.

Signed-off-by: Federico Vaga <federicov@linino.org>